ipq806x: 5.4: update nss-volt
QCA QSDK driver has a issue that it will probe too early, in fact so early that the requested regulator will not even be registered.
Regulator core will then return a dummy regulator instead, this does not suit us a then driver will panic because you cant get/set voltage on a dummy regulator.
That will also cause severe NSS core instability due to being unable to scale their voltage with frequency.

So lets update the driver to check whether returned regulator is a dummy one, if so simply deffer the driver probe.
This way the regulator core will call the driver probe again later when the real regulator will be registered.

Also update the driver prints, name and move it to the regulator directory where it belongs.
Export the needed scaling function as a header and add a Kconfig symbol instead of building along the GCC clock driver.

Enable it by default in kernel config as its needed for NSS.

Signed-off-by: Robert Marko <robimarko@gmail.com>