ipq807x: add DVFS support
This patchset adds support for DVFS on the IPQ8074 family.

Its split into HK and AC CPU DTSI as they have different clocks and
voltages.
Currently, a static voltage map is used instead of the CPR as there is no
driver for it to dynamically adjust each range based on the load.

Signed-off-by: Robert Marko <robimarko@gmail.com>