ipq807x: set network interface names
Set network interface names according to physical port markings.

Since this would break networking configs, compat version is set for
affected boards along with a message when trying to sysupgrade that it
can only be done if not preserving config.

Compat version will be dropped once a PR to OpenWrt is made.

Signed-off-by: Robert Marko <robimarko@gmail.com>